减少名称污染
